requirements.txt
streamlit
requests
langdetect
langchain
openai==0.28
anthropic
ollama
python-dotenv
FPDF